+++ apps/expo-go/android/expoview/src/main/java/host/exp/exponent/taskManager/ExpoHeadlessAppLoader.kt
@@ -36,21 +36,18 @@ class ExpoHeadlessAppLoader @DoNotStrip constructor(context: Context?) : Headles
         )
         val appRecord = appLoader.loadApp(
           params.appUrl,
-          mapOf(),
-          object : AppLoaderProvider.Callback() {
-            override fun onComplete(success: Boolean, exception: Exception?) {
-              if (exception != null) {
-                exception.printStackTrace()
-                Log.e(TAG, exception.message!!)
-              }
-              HeadlessAppLoaderNotifier.notifyAppLoaded(params.appScopeKey)
-              callback.apply(success)
-              if (!success) {
-                appScopeKeysToAppRecords.remove(params.appScopeKey)
-              }
-            }
+          mapOf()
+        ) { success, exception ->
+          if (exception != null) {
+            exception.printStackTrace()
+            Log.e(TAG, exception.message!!)
           }
-        )
+          HeadlessAppLoaderNotifier.notifyAppLoaded(params.appScopeKey)
+          callback.apply(success)
+          if (!success) {
+            appScopeKeysToAppRecords.remove(params.appScopeKey)
+          }
+        }
 
         appScopeKeysToAppRecords[params.appScopeKey] = appRecord
       }