+++ packages/expo-modules-core/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 💡 Others
 
+- [Android] Migrate `AppLoaderProvider` to kotlin.
+
 ## 2.3.1 — 2025-04-08
 
 _This version does not introduce any user-facing changes._